Update README.md
Recycling capstone project 
 
At Queens university of charlotte Even though we have separate trash and recycling bins, College students often times do not separate their waste properly. Because of this, many universities end up throwing everything into the landfill; worsening Global Warming. We are working on a machine learning project with a focus on python as the coding language. we were tasked with solving this recycling problem at Queens University of Charlotte. Using python and machine learning, we are attempting to train a raspberry pi3 to distinguish the difference in waste materials. I will then code responses for each material (i.e. plastic vs paper), and then distinguish which materials can be recycled and which cannot. My hope is to create a way recycle in the 21st century. 
 
 
What is the problem? 
 
Even though we have separate trash and recycling bins, College students often times do not separate their waste properly. Because of this, many universities end up throwing everything into the landfill; worsening Global Warming. 
 
Why is it important? 
 
To combat Global Warming, we need to switch over to more sustainable lifestyles, and that includes properly sorting our waste from our recycling. Without an improved way most universities are instead sending everything to the landfill. Unfortunately, this is the case at many universities throughout America, and Queens is no exception. 
 
How are we doing this?  
 
We have designed a machine that looks at an item and, using image recognition, determines whether the item is trash, recyclable, or compostable, and then sorts it accordingly. 
 
What is machine learning? 
 
Machine learning is an application of artificial intelligence (AI) that provides a system the ability to automatically learn and improve from experience without being furthered programmed. Machine learning focuses on the development of computer programs that can access data and use it to learn for themselves. Machine learning, learns as more and more data gets inputted or collected.